+++ src/main/java/net/frozenblock/freezeframe/registry/FFStats.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2026 FrozenBlock
+ * This file is part of Freeze Frame.
+ *
+ * This program is free software; you can modify it under
+ * the terms of version 1 of the FrozenBlock Modding Oasis License
+ * as published by FrozenBlock Modding Oasis.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * FrozenBlock Modding Oasis License for more details.
+ *
+ * You should have received a copy of the FrozenBlock Modding Oasis License
+ * along with this program; if not, see <https://github.com/FrozenBlock/Licenses>.
+ */
+
+package net.frozenblock.freezeframe.registry;
+
+import net.frozenblock.freezeframe.FFConstants;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.Identifier;
+import net.minecraft.stats.StatFormatter;
+import net.minecraft.stats.Stats;
+
+public class FFStats {
+	public static final Identifier INTERACT_WITH_DEVELOPING_TABLE = makeCustomStat("interact_with_developing_table", StatFormatter.DEFAULT);
+
+	public static void init() {}
+
+	private static Identifier makeCustomStat(String name, StatFormatter formatter) {
+		final Identifier id = FFConstants.id(name);
+		Registry.register(BuiltInRegistries.CUSTOM_STAT, id, id);
+		Stats.CUSTOM.get(id, formatter);
+		return id;
+	}
+}