+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
+- Bumped Freeze Frame's protocol version to 2.
+- Added a statistic for interact with a Developing Table.
 - Added a config option to render Photo items in third-person, similarly to their first-person counterpart.
 - Photo rendering can now render a backside.
   - Added the `FRAME_BACK` Frame Type.